Create stairSteps.py
allPaths() returns every combination of possible steps up a staircase of any length. For instance, if there is a staircase of length 3    
(3 steps), and you can either take 1 or 2 steps at a time, calling allPaths() on this staircase would return [ [1,1,1], [1,2], [2,1] ].

The final for-loop then calculates the sum of the associated staircase values and returns the max.